requirements.txt
torch
torchvision
transformers
accelerate
Pillow
matplotlib
numpy
qwen-vl-utils